[lein] Add first Clojure leiningen profile.
This is my first attempt at a configuration file for my use of the
Clojure build, package, and dependency manager tool of choice
leiningen.  So far, this just involves pulling in lein-try to try
different sample packages with explicitly setting up a repository with a
deps manifest.  As the name implies, it is for experimentation.

https://github.com/rkneufeld/lein-try