[.xinitrc] Add proper variable for Java GUI apps.
There are issues with the tiling window managers I prefer.

You must add the necessary _JAVA_AWT_WM_NONPARENTING appropriately or
Swing and other Java GUI apps will occasionally show up as just grey
background boxes with the UI missing.  This is frequently the case with
Android Studio at the time of this writing with the ratpoison WM.